# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ProjectOps API",
    "description": "Autonomous multi-agent project execution and risk management system. Deterministic scheduling, immutable versioned plans, and append-only audits.",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/projects": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "List Projects",
        "description": "List all projects.",
        "operationId": "list_projects_api_v1_projects_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProjectResponse"
                  },
                  "type": "array",
                  "title": "Response List Projects Api V1 Projects Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Create Project",
        "description": "Create a new project.",
        "operationId": "create_project_api_v1_projects_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "Get Project",
        "description": "Get project details including members.",
        "operationId": "get_project_api_v1_projects__project_id__get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/members": {
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Add Project Member",
        "description": "Add a team member to a project.",
        "operationId": "add_project_member_api_v1_projects__project_id__members_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemberCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "List Project Members",
        "description": "List members assigned to a project.",
        "operationId": "list_project_members_api_v1_projects__project_id__members_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MemberResponse"
                  },
                  "title": "Response List Project Members Api V1 Projects  Project Id  Members Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/plans": {
      "post": {
        "tags": [
          "plans"
        ],
        "summary": "Create Plan Snapshot",
        "description": "Submit a new immutable plan snapshot (Rule 5: version increment, no overwrite).",
        "operationId": "create_plan_snapshot_api_v1_projects__project_id__plans_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanResponseSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/plans/active": {
      "get": {
        "tags": [
          "plans"
        ],
        "summary": "Get Active Plan",
        "description": "Fetch the currently active plan snapshot.",
        "operationId": "get_active_plan_api_v1_projects__project_id__plans_active_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanResponseSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/plans/versions": {
      "get": {
        "tags": [
          "plans"
        ],
        "summary": "List Plan Versions",
        "description": "List all immutable version numbers available for a project.",
        "operationId": "list_plan_versions_api_v1_projects__project_id__plans_versions_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanVersionListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/plans/{version}": {
      "get": {
        "tags": [
          "plans"
        ],
        "summary": "Get Plan By Version",
        "description": "Retrieve an immutable historical plan snapshot by its version number.",
        "operationId": "get_plan_by_version_api_v1_projects__project_id__plans__version__get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Version"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanResponseSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DependencySchema": {
        "properties": {
          "predecessor_id": {
            "type": "string",
            "title": "Predecessor Id"
          },
          "successor_id": {
            "type": "string",
            "title": "Successor Id"
          },
          "dep_type": {
            "$ref": "#/components/schemas/DependencyType",
            "default": "FINISH_TO_START"
          },
          "lag_days": {
            "type": "integer",
            "title": "Lag Days",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "predecessor_id",
          "successor_id"
        ],
        "title": "DependencySchema",
        "description": "Precedence dependency relationship between tasks."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "MemberCreateRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name"
          },
          "role": {
            "type": "string",
            "maxLength": 100,
            "title": "Role",
            "default": "engineer"
          },
          "daily_capacity_hours": {
            "type": "number",
            "maximum": 24.0,
            "exclusiveMinimum": 0.0,
            "title": "Daily Capacity Hours",
            "default": 8.0
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "MemberCreateRequest",
        "description": "Payload to add a member to a project."
      },
      "MemberResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "daily_capacity_hours": {
            "type": "number",
            "title": "Daily Capacity Hours"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "name",
          "role",
          "daily_capacity_hours",
          "created_at"
        ],
        "title": "MemberResponse",
        "description": "Response schema for a project member."
      },
...
      "PlanCreateRequest": {
        "properties": {
          "version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Version",
            "description": "Monotonically increasing version number"
          },
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name"
          },
          "target_completion_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Completion Date"
          },
          "created_by": {
            "type": "string",
            "maxLength": 255,
            "title": "Created By",
            "default": "system"
          },
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/TaskCreateSchema"
            },
            "type": "array",
            "title": "Tasks"
          },
          "dependencies": {
            "items": {
              "$ref": "#/components/schemas/DependencySchema"
            },
            "type": "array",
            "title": "Dependencies"
          }
        },
        "type": "object",
        "required": [
          "version",
          "name"
        ],
        "title": "PlanCreateRequest",
        "description": "Payload to create an immutable plan snapshot."
      },
      "PlanResponseSchema": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "target_completion_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Completion Date"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "created_by": {
            "type": "string",
            "title": "Created By"
          },
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/TaskResponseSchema"
            },
            "type": "array",
            "title": "Tasks",
            "default": []
          },
          "dependencies": {
            "items": {
              "$ref": "#/components/schemas/DependencySchema"
            },
            "type": "array",
            "title": "Dependencies",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "version",
          "name",
          "target_completion_date",
          "created_at",
          "created_by"
        ],
        "title": "PlanResponseSchema",
        "description": "Response schema for a complete versioned plan snapshot."
      },
      "PlanVersionListResponse": {
        "properties": {
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "versions": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Versions"
          },
          "active_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Version"
          }
        },
        "type": "object",
        "required": [
          "project_id",
          "versions",
          "active_version"
        ],
        "title": "PlanVersionListResponse",
        "description": "Response schema listing available plan versions for a project."
      },
...
      "ProjectCreateRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "type": "string",
            "maxLength": 2000,
            "title": "Description",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "ProjectCreateRequest",
        "description": "Payload to create a new project."
      },
      "ProjectResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "members": {
            "items": {
              "$ref": "#/components/schemas/MemberResponse"
            },
            "type": "array",
            "title": "Members",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "created_at",
          "updated_at"
        ],
        "title": "ProjectResponse",
        "description": "Response schema for a project."
      },
...
      "TaskCreateSchema": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Id"
          },
          "title": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Title"
          },
          "description": {
            "type": "string",
            "maxLength": 2000,
            "title": "Description",
            "default": ""
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus",
            "default": "TODO"
          },
          "estimated_hours": {
            "type": "number",
            "minimum": 0.0,
            "title": "Estimated Hours",
            "default": 8.0
          },
          "assigned_to_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned To Id"
          },
          "due_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Date"
          },
          "start_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title"
        ],
        "title": "TaskCreateSchema",
        "description": "Schema for defining a task when creating or updating a plan snapshot."
      },
      "TaskResponseSchema": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus"
          },
          "estimated_hours": {
            "type": "number",
            "title": "Estimated Hours"
          },
          "assigned_to_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned To Id"
          },
          "due_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Date"
          },
          "start_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "is_overdue": {
            "type": "boolean",
            "title": "Is Overdue",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "description",
          "status",
          "estimated_hours",
          "assigned_to_id",
          "due_date",
          "start_date",
          "completed_at"
        ],
        "title": "TaskResponseSchema",
        "description": "Schema for task representation in a plan snapshot response."
      },
      "TaskStatus": {
        "type": "string",
        "enum": [
          "TODO",
          "IN_PROGRESS",
          "BLOCKED",
          "COMPLETED"
        ],
        "title": "TaskStatus",
        "description": "Valid task statuses.\nRule: OVERDUE is strictly a derived property, never a stored status."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}